requirements.txt
Django==3.2
django_admin_rangefilter==0.9.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-extensions==3.2.1
django_import_export==3.0.2
django_rest_knox==4.2.0
djangorestframework==3.12.4
django-rest-swagger==2.2.0
drf_yasg==1.20.0
import_export==0.3.1
knox==0.1.14
paho_mqtt==1.6.1
Pillow==9.4.0
python-dateutil==2.8.2
requests==2.22.0
